pyproject.toml
[tool]
[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ['F', 'E', 'W', 'B', 'I']
ignore = [
    'E402',

    # multiple statements on one line. Incompatible with:
    # if cond: break
    'E701',

    # syntax error in forward declaration - incompatible with jaxtyping
    'F722',
]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
include = ['.']
exclude = [
    '.venv/',
    'sweep/',
    'scripts/',
    'build/',
    'corerl/environment/smpl',
    'corerl/environment/third_party',
    'corerl/sql_logging/examples',
]
venvPath = '.'
venv = '.venv'
typeCheckingMode = 'standard'

# common overrides
useLibraryCodeForTypes = true
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
enableReachabilityAnalysis = true

# stricter overrides
reportInconsistentConstructor = true
reportMissingParameterType = true
reportUnnecessaryCast = true
reportUnnecessaryIsInstance = true
reportUnusedClass = true
reportUnusedExpression = true
reportUnusedFunction = true


[tool.deadcode]
exclude = [".venv", "docs", "test", "build", "corerl/environment/smpl"]
ignore-names = ["_*"]
ignore-names-in-files = []

[tool.commitizen]
name = "cz_conventional_commits"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true

[tool.coverage.run]
omit = ["test/*"]
concurrency = ["thread", "multiprocessing"]

# If the vendor directory exists and contains linesearchopt dependency
# this will be included in our generated wheel. Otherwise linesearchopt
# will need to be installed manually prior to corerl. This is necessary
# to support our docker image for running corerl.
[tool.setuptools.packages.find]
where = [".", "vendor"]
include = ["corerl*", "linesearchopt*"]

[tool.setuptools.package-dir]
"linesearchopt" = "vendor/linesearchopt"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "corerl"
version = "0.26.2"
description = ""
requires-python = ">=3.12,<3.13"
authors = []
dependencies = [
    "asyncua~=1.0",
    "boto3~=1.35",
    "casadi~=3.6",
    "cloudpathlib[s3]~=0.20",
    "gymnasium~=1.0",
    "influxdb_client~=1.39",
    "jaxtyping~=0.2",
    "matplotlib~=3.8",
    "nltk~=3.8",
    "numba~=0.60",
    "numpy~=1.26",
    "pandas~=2.2",
    "pydantic~=2.9",
    "pygame~=2.5",
    "pyyaml~=6.0",
    "requests~=2.31",
    "scipy~=1.13",
    "sqlalchemy~=2.0",
    "sqlalchemy-utils~=0.41",
    "textual~=1.0.0",
    "torch~=2.5",
    "tqdm~=4.66",
    "scikit-learn~=1.5",
    "websockets~=14.1",
    "psycopg2-binary~=2.9",
]
readme = "README.md"

[project.optional-dependencies]
linesearchopt_gh = [
    "linesearchopt@git+ssh://git@github.com/rlcoretech/linesearchopt.git",
]

dev = [
    "pip",
    "uv",
    "docker",
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pytest-coverage",
    "pytest-xdist",
    "pytest-timeout",
    "pyright",
    "debugpy",
    "boto3",
    "linesearchopt@git+ssh://git@github.com/rlcoretech/linesearchopt.git",
]

[project.scripts]
corerl_main = "corerl.main:main"
corerl_convertdb_backup = "corerl.bin.convert_db_backup:main"
corerl_opc_explorer = "corerl.bin.opc_explorer:main"

